refactor(codec): support enum opcodes in PacketOpcodeAttribute
- Add constructor overload accepting enum values, validated and
  converted to ushort at compile-time-safe attribute usage
- Remove Bytes32Formatter and PacketHeaderFormatter (redundant
  with generic unmanaged serialization path)
- Remove static constructor from LiteSerializer that registered
  the deleted formatters
- Use enum directly in [PacketOpcode] across all handlers instead
  of manual (ushort) casts
- Add exhaustive ControlType switch cases in SystemTimeSyncHandlers
- Remove unused usings and trailing whitespace

+++ src/Nalix.Runtime/Handlers/HandshakeHandlers.cs
@@ -40,7 +40,7 @@ public sealed class HandshakeHandlers
     [ReservedOpcodePermitted]
     [PacketEncryption(false)]
     [PacketPermission(PermissionLevel.NONE)]
-    [PacketOpcode((ushort)ProtocolOpCode.SESSION_INIT)]
+    [PacketOpcode(ProtocolOpCode.SESSION_INIT)]
     public static async ValueTask HandleSessionInitAsync(IPacketContext<SessionInit> context)
     {
         ArgumentNullException.ThrowIfNull(context);
